refactor: storageKeys.ts constant로 옮기고 네이밍 변경
